fix(materials): address Copilot review feedback on PR #130
Three Copilot comments all worth applying:

1. React.ReactNode typed without importing React

   The viewer declared `let rendered: React.ReactNode` but never imported
   React, relying on an ambient global that isn't guaranteed under the
   repo's TS setup. Import `type { ReactNode } from 'react'` and use the
   unqualified name — consistent with how ReactNode is referenced
   elsewhere in the codebase.

2. decodeURIComponent can throw URIError on malformed encoding

   A hand-edited URL (lone `%`, invalid UTF-8 sequence) causes
   `decodeURIComponent` to throw and 500 the page instead of rendering a
   404. Added a small `safeDecode` helper that falls back to the raw
   value on decode failure; the downstream findMarketingAsset lookup
   will then naturally miss and trigger notFound(). Matches the pattern
   other routes in the repo use.

3. Viewer reads asset.filePath directly instead of going through the
   path-allowlist + realpath check used by the raw asset API route

   The real fix: extract the safe-read helper from the API handler into
   a shared backend/marketing/asset-read.ts module so every caller that
   reads from a runtime-doc-controlled path gets the same guarantees.
   Viewer + handler both now call readMarketingAssetWithinAllowedRoots;
   the 85 lines of path normalization / trusted-root allowlist / realpath
   symlink check live in exactly one place.

   Runtime asset paths originate server-side but can be influenced by
   tenant input (pipeline output, uploads), so treating them as
   untrusted and funneling every read through the allowlist helper is
   the right default. Asset-library.ts's findMarketingAsset stays
   metadata-only; the actual file read is isolated to asset-read.ts.

Verification: npm run typecheck clean; simple-markdown +
brand-review-output + runtime-views-auth-hardening 29/29 pass.
