+++ backend/marketing/asset-read.ts
@@ -1,0 +1,150 @@
+import { readFile, realpath } from 'node:fs/promises';
+import { tmpdir } from 'node:os';
+import path from 'node:path';
+
+import { resolveCodePath, resolveCodeRoot, resolveDataRoot } from '@/lib/runtime-paths';
+
+/**
+ * Safe reader for files referenced by the marketing runtime documents.
+ *
+ * Every path that lands on disk through the marketing pipeline eventually
+ * needs to be read back out (API handler serving raw bytes, materials
+ * viewer rendering markdown, preview endpoints). We centralize the read
+ * so both callers get the same guarantees:
+ *
+ *   1. `normalizeAssetPath` strips any `.` / `..` segments so relative
+ *      paths cannot climb out of the trusted root set they'll be joined
+ *      against.
+ *   2. `trustedRoots` is the closed set of directories we're willing to
+ *      serve files from — DATA_ROOT (runtime artifacts), CODE_ROOT (repo
+ *      source), Lobster CWD/stage caches (generated pipeline outputs).
+ *   3. `isWithinRoot` both before and after `realpath` resolution ensures
+ *      symlink traversal doesn't escape the trusted root (an attacker who
+ *      could plant a symlink inside the root still can't use it to read
+ *      an arbitrary file outside it).
+ *   4. `absoluteCompatibilityCandidates` handles the legacy `/app/aries-app`
+ *      → `/app` path rewrite without widening the trust boundary.
+ *
+ * Returns the file contents as a Buffer when the path was safe and the
+ * file existed, otherwise `null`. Throws on unexpected I/O errors so the
+ * caller can decide whether to bubble up a 500.
+ *
+ * Runtime asset paths originate server-side (generated by the pipeline
+ * or written by the onboarding/upload flows), but since those same flows
+ * can be influenced by tenant input we treat them as untrusted and funnel
+ * every read through this helper instead of calling `readFile` directly.
+ */
+export async function readMarketingAssetWithinAllowedRoots(
+  filePath: string,
+): Promise<Buffer | null> {
+  const normalizedPath = normalizeAssetPath(filePath);
+  if (!normalizedPath) {
+    return null;
+  }
+
+  const roots = trustedRoots();
+  const candidates =
+    normalizedPath.kind === 'absolute'
+      ? absoluteCompatibilityCandidates(normalizedPath.path)
+      : roots.map((root) => path.resolve(root, normalizedPath.path));
+
+  for (const candidate of candidates) {
+    for (const root of roots) {
+      if (!isWithinRoot(root, candidate)) {
+        continue;
+      }
+      try {
+        const [resolvedRoot, resolvedCandidate] = await Promise.all([
+          realpath(root).catch(() => root),
+          realpath(candidate),
+        ]);
+        if (!isWithinRoot(resolvedRoot, resolvedCandidate)) {
+          continue;
+        }
+        return await readFile(resolvedCandidate);
+      } catch (error) {
+        if (
+          error &&
+          typeof error === 'object' &&
+          'code' in error &&
+          (error.code === 'ENOENT' || error.code === 'ENOTDIR')
+        ) {
+          continue;
+        }
+        throw error;
+      }
+    }
+  }
+
+  return null;
+}
+
+type NormalizedAssetPath =
+  | { kind: 'relative'; path: string }
+  | { kind: 'absolute'; path: string };
+
+function normalizeAssetPath(filePath: string): NormalizedAssetPath | null {
+  const trimmed = filePath.trim();
+  if (!trimmed) {
+    return null;
+  }
+
+  if (path.isAbsolute(trimmed)) {
+    return {
+      kind: 'absolute',
+      path: path.normalize(trimmed),
+    };
+  }
+
+  const segments = trimmed.split(/[\\/]+/).filter(Boolean);
+  if (segments.length === 0 || segments.some((segment) => segment === '.' || segment === '..')) {
+    return null;
+  }
+
+  return {
+    kind: 'relative',
+    path: path.join(...segments),
+  };
+}
+
+function trustedRoots(): string[] {
+  return Array.from(
+    new Set(
+      [
+        resolveDataRoot(),
+        resolveCodeRoot(),
+        resolveCodePath('lobster'),
+        process.env.OPENCLAW_LOCAL_LOBSTER_CWD?.trim(),
+        process.env.OPENCLAW_LOBSTER_CWD?.trim(),
+        process.env.LOBSTER_STAGE1_CACHE_DIR?.trim() || path.join(tmpdir(), 'lobster-stage1-cache'),
+        process.env.LOBSTER_STAGE2_CACHE_DIR?.trim() || path.join(tmpdir(), 'lobster-stage2-cache'),
+        process.env.LOBSTER_STAGE3_CACHE_DIR?.trim() || path.join(tmpdir(), 'lobster-stage3-cache'),
+        process.env.LOBSTER_STAGE4_CACHE_DIR?.trim() || path.join(tmpdir(), 'lobster-stage4-cache'),
+      ].filter((value): value is string => typeof value === 'string' && value.trim().length > 0),
+    ),
+  );
+}
+
+function absoluteCompatibilityCandidates(filePath: string): string[] {
+  const normalized = path.normalize(filePath);
+  const candidates = new Set([normalized]);
+  const codeRoot = path.normalize(resolveCodeRoot());
+  const legacyCodeRoot = path.join(codeRoot, 'aries-app');
+
+  if (normalized === legacyCodeRoot || normalized.startsWith(`${legacyCodeRoot}${path.sep}`)) {
+    const suffix = normalized.slice(legacyCodeRoot.length);
+    candidates.add(path.join(codeRoot, suffix));
+  }
+
+  return Array.from(candidates);
+}
+
+/**
+ * Returns true only when the resolved candidate stays within the provided
+ * root directory. Paths that escape upward or resolve to another absolute
+ * location are rejected.
+ */
+function isWithinRoot(root: string, candidate: string): boolean {
+  const relative = path.relative(root, candidate);
+  return relative === '' || (!relative.startsWith('..') && !path.isAbsolute(relative));
+}
